# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniCreator AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: List user's organizations
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationBody'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
  /organizations/{orgId}:
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Get organization details
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Update organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationBody'
      responses:
        '200':
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'

  # WORKSPACES
...
components:
...
  schemas:
...
    Organization:
      type: object
      required: [id, name, slug, plan, ownerId, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        plan:
          type: string
        ownerId:
          type: string
        creditsBalance:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateOrganizationBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: ["string", "null"]

    # WORKSPACES
...